+++ carla_ros_bridge/src/carla_ros_bridge/walker.py
@@ -73,9 +73,9 @@ class Walker(Actor):
         :rtpye : std_msgs.msg.ColorRGBA
         """
         color = ColorRGBA()
-        color.r = 255
+        color.r = 0
         color.g = 0
-        color.b = 0
+        color.b = 255
         return color
 
     def get_marker(self):
@@ -111,7 +111,7 @@ class Walker(Actor):
         marker.scale.x = self.carla_actor.bounding_box.extent.x * 2.0
         marker.scale.y = self.carla_actor.bounding_box.extent.y * 2.0
         marker.scale.z = self.carla_actor.bounding_box.extent.z * 2.0
-        self.publish_message('/carla/walker_marker', marker)
+        self.publish_message('/carla/marker', marker)
 
     def get_object_info(self):
         """